requirements.txt
streamlit
psycopg2-binary
python-dotenv
google-generativeai
streamlit-folium
folium